+++ packages/elements/src/photon-multirx-form/prescribe-workflow-analytics.test.tsx
@@ -309,6 +309,7 @@ function renderPrescribeWorkflow(
         HttpResponse.json({
           data: {
             me: {
+              id: 'prov_1',
               signatureAttestationStatus: {
                 __typename: 'NeedsSignatureAttestation',
                 version: 'v1',
